requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attributedict==0.3.0
attrs==23.2.0
autoawq==0.2.0
autoawq_kernels==0.0.6
blessings==1.7
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
coverage==7.4.4
DataProperty==1.0.1
datasets==2.14.7
deepdiff==7.0.1
dill==0.3.7
distlib==0.3.8
evaluate==0.4.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
inspecta==0.1.3
Jinja2==3.1.3
joblib==1.4.0
jsonlines==4.0.0
lm_eval==0.4.1
lxml==5.2.1
MarkupSafe==2.1.5
mbstrdecoder==1.1.3
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
networkx==3.2.1
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
ordered-set==4.1.0
packaging==24.0
pandas==2.2.2
pathvalidate==3.2.0
peft==0.10.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.5.0
portalocker==2.8.2
protobuf==5.26.1
psutil==5.9.8
pyarrow==16.0.0
pyarrow-hotfix==0.6
pybind11==2.12.0
Pygments==2.17.2
pyproject-api==1.6.1
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
responses==0.18.0
rootpath==0.1.1
rouge_score==0.1.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
six==1.16.0
sqlitedict==2.1.0
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.4.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tox==4.14.2
tqdm==4.66.2
tqdm-multiprocess==0.0.11
transformers==4.40.0
triton==2.2.0
typepy==1.3.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.3
word2number==1.1
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0
